services/orchestrator/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
grpcio==1.64.0
protobuf==5.27.0
torch==2.3.0
httpx==0.27.0
minio==7.2.7
pydantic==2.7.0
python-jose[cryptography]==3.3.0
cryptography==42.0.8
requests==2.32.3
